rename: hook 위치 hooks로 수정
