--- src/hook/useNavigateToPrev.ts
@@ -1,9 +1,0 @@
-import { useNavigate } from 'react-router-dom';
-
-const useNavigateToPrev = () => {
-  const navigate = useNavigate();
-
-  return () => navigate(-1);
-};
-
-export default useNavigateToPrev;
